feat: Add baml dynamic typing
