+++ cognee/infrastructure/llm/extraction/extract_event_entities.py
@@ -1,6 +1,7 @@
 import os
-from typing import List, Type
+from typing import Type
 from pydantic import BaseModel
+from cognee.infrastructure.llm.prompts.render_prompt import render_prompt
 from cognee.infrastructure.llm.LLMGateway import LLMGateway
 from cognee.infrastructure.llm.config import (
     get_llm_config,
@@ -35,7 +36,7 @@ async def extract_event_entities(content: str, response_model: Type[BaseModel]):
     else:
         base_directory = None
 
-    system_prompt = LLMGateway.render_prompt(prompt_path, {}, base_directory=base_directory)
+    system_prompt = render_prompt(prompt_path, {}, base_directory=base_directory)
 
     content_graph = await LLMGateway.acreate_structured_output(
         content, system_prompt, response_model
