+++ cognee/infrastructure/llm/extraction/knowledge_graph/extract_content_graph.py
@@ -2,6 +2,7 @@ import os
 from typing import Type, Optional
 from pydantic import BaseModel
 
+from cognee.infrastructure.llm.prompts import render_prompt
 from cognee.infrastructure.llm.LLMGateway import LLMGateway
 from cognee.infrastructure.llm.config import (
     get_llm_config,
@@ -26,7 +27,7 @@ async def extract_content_graph(
         else:
             base_directory = None
 
-        system_prompt = LLMGateway.render_prompt(prompt_path, {}, base_directory=base_directory)
+        system_prompt = render_prompt(prompt_path, {}, base_directory=base_directory)
 
     content_graph = await LLMGateway.acreate_structured_output(
         content, system_prompt, response_model
